refactor

+++ src/server.js
@@ -91,64 +91,34 @@ app.get('/auth/logout', (req, res) => {
 });
 
 app.post('/spreadsheets', async (req, res) => {
-  //res.status(500).json({ error: 'ololoshechki' });
   const { token, refreshToken, estimateId, userId } = req.body;
   const credentials = { access_token: token, refresh_token: refreshToken };
   const spHelper = spreadSheets(credentials);
   const estimate = await Estimate.findById(estimateId);
-  // probably estimate should be retrieved via grahpql
-  // if so - this piece of code should be replaced
-  if (!estimate.spreadsheetId || !estimate.spreadsheetId[userId]) {
-    spHelper.createSpreadsheet(estimate, async (err, sp) => {
-      if (err) {
-        // if access_token has been expired
-        // it should be recreated by refreshToken
-        if (err.code == 401) {
-          const newCredentials = await spHelper.updateCredentials();
-          if (newCredentials) {
-            const query = { 'google.token': token };
-            const projection = { $set: { google: newCredentials } };
-            User.update(query, projection);
-          }
-        }
-        res.status(400).send({ error: true, message: err });
-        res.end();
-      }
-      const { spreadsheetId } = sp;
-      const estId = await Estimate.update({ _id: estimateId}, { $set: { [`spreadsheetId.${userId}`]: spreadsheetId } });
-      res.status(200).send({ message: `Spreadsheet ${spreadsheetId} updated` });
-      res.end();
-    });
-  } else {
-
+  if (estimate.spreadsheetId && estimate.spreadsheetId[userId]) {
     const file = await spHelper.getFile(estimate.spreadsheetId[userId]);
-    console.log('fileId');
-    console.log(file);
     if (file && file.id) {
       await spHelper.deleteFile(file.id);
     }
-    spHelper.createSpreadsheet(estimate, async (err, sp) => {
-      if (err) {
-        // if access_token has been expired
-        // it should be recreated by refreshToken
-        if (err.code == 401) {
-          const newCredentials = await spHelper.updateCredentials();
-          if (newCredentials) {
-            const query = { 'google.token': token };
-            const projection = { $set: { google: newCredentials } };
-            User.update(query, projection);
-          }
-        }
-        res.status(400).send({ error: true, message: err });
-        res.end();
-      }
-      const { spreadsheetId } = sp;
-      const estId = await Estimate.update({ _id: estimateId}, { $set: { [`spreadsheetId.${userId}`]: spreadsheetId } });
-      res.status(200).send({ message: `Spreadsheet ${spreadsheetId} updated` });
-      res.end();
-    });
-
   }
+  spHelper.createSpreadsheet(estimate, async (err, sp) => {
+    if (err) {
+      if (err.code == 401) {
+        const newCredentials = await spHelper.updateCredentials();
+        if (newCredentials) {
+          const query = { 'google.token': token };
+          const projection = { $set: { google: newCredentials } };
+          User.update(query, projection);
+        }
+      }
+      res.status(400).send({ error: true, message: err });
+      res.end();
+    }
+    const { spreadsheetId } = sp;
+    const estId = await Estimate.update({ _id: estimateId}, { $set: { [`spreadsheetId.${userId}`]: spreadsheetId } });
+    res.status(200).send({ message: `Spreadsheet ${spreadsheetId} updated` });
+    res.end();
+  });
 });
 
 //
